win,fs: Continue to set `FILE_ATTRIBUTE_READONLY`
When switching to use ACLs, we dropped setting the "readonly" filesystem
bit.  With this change, we continue to set the "readonly" bit if there
are no write permissions assigned to within the permissions mask.  This
allows `stat()` to maintain its half-functional state, returning `0666`
if the file is writable and `0444` otherwise.